Move color_parser.dart inside data folder since it's used only there
